requirements.txt
RPi.GPIO
pytesseract
tesseract
matplotlib
requests
imutils
opencv-python